collision controler + other small fixes + tehnical commit
trying out a non nav mesh enemy movement / attacking.
first steppts towards that

+++ Rogue, The Cube/Assets/Scripts/Enemy/Template/Enemy.cs
@@ -41,8 +41,8 @@ public class Enemy : ScriptableObject
 
     [Header("Enemy AI")]
     public EnemyType enemyType;
-    public int detectionSphereRadius;
-    public int attackingDistance;
+    public float detectionSphereRadius;
+    public float attackingDistance;
     [Header("Stamina Gain")]
     public float staminaEverySeconds;
     public int staminGainAmmount;
